requirements.txt
python-dotenv==1.0.1
sqlalchemy==2.0.38
PyMySQL==1.1.1
cryptography==43.0.3
#-e .